# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Unsaid API
  version: 0.1.0
  description: Backend for Unsaid, a communication aid for people with non-fluent aphasia. Not a medical device; makes no diagnostic or therapeutic claims. Generated from zod schemas (src/http/schemas.ts) by `pnpm openapi`.
  license:
    name: MIT
    identifier: MIT
servers:
  - url: http://localhost:8080
    description: Local
...
paths:
...
  /v1/omi/status:
    get:
      tags:
        - Omi
      summary: Is the Omi live path delivering?
      description: Last segment time, source (OMI_REALTIME vs SIMULATED) and 5-minute counts.
      operationId: get_v1_omi_status
      parameters:
        - name: userId
          in: query
          required: false
          schema:
            type: string
            minLength: 1
            maxLength: 128
      security:
        - sessionCookie: []
        - bearerAuth: []
      responses:
        "200":
          description: Status
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    const: true
                  data:
                    $ref: "#/components/schemas/OmiStatus"
                required:
                  - ok
                  - data
                additionalProperties: false
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: Missing or invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "429":
          description: Rate limited
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /v1/segments:
    get:
      tags:
        - Omi
      summary: Transcript segments (recent, or by dedupe key for fact provenance)
      operationId: get_v1_segments
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 128
        - name: ids
          in: query
          required: false
          schema:
            description: Comma-separated dedupe keys (max 50). Omit to list the most recent segments.
            type: string
            maxLength: 4000
        - name: limit
          in: query
          required: false
          schema:
            default: 50
            type: integer
            minimum: 1
            maximum: 200
      security:
        - sessionCookie: []
        - bearerAuth: []
      responses:
        "200":
          description: Segments, oldest first
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    const: true
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Segment"
                required:
                  - ok
                  - data
                additionalProperties: false
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: Missing or invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "429":
          description: Rate limited
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
components:
  schemas:
    ErrorEnvelope:
      $id: "#/components/schemas/ErrorEnvelope"
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              example: validation_error
            message:
              type: string
            requestId:
              type: string
            details: {}
          required:
            - code
            - message
          additionalProperties: false
      required:
        - error
      additionalProperties: false
...
    OmiStatus:
      $id: "#/components/schemas/OmiStatus"
      type: object
      properties:
        lastSegmentAt:
          anyOf:
            - type: string
              format: date-time
            - type: "null"
        lastSegmentSource:
          anyOf:
            - type: string
              enum:
                - OMI_REALTIME
                - OMI_MEMORY
                - SIMULATED
            - type: "null"
        segmentsLast5Min:
          type: object
          properties:
            OMI_REALTIME:
              type: integer
            OMI_MEMORY:
              type: integer
            SIMULATED:
              type: integer
          required:
            - OMI_REALTIME
            - OMI_MEMORY
            - SIMULATED
          additionalProperties: false
        lastRawWebhookAt:
          anyOf:
            - type: string
              format: date-time
            - type: "null"
        rawWebhooksLast5Min:
          type: integer
        webhookSecretConfigured:
          type: boolean
      required:
        - lastSegmentAt
        - lastSegmentSource
        - segmentsLast5Min
        - lastRawWebhookAt
        - rawWebhooksLast5Min
        - webhookSecretConfigured
      additionalProperties: false
...
    Segment:
      $id: "#/components/schemas/Segment"
      type: object
      properties:
        id:
          type: string
          minLength: 1
          maxLength: 128
        dedupeKey:
          type: string
          description: Matches `sourceSegmentIds` on memory facts
        sessionId:
          type: string
        text:
          type: string
        speaker:
          type:
            - string
            - "null"
        isUser:
          type: boolean
        source:
          type: string
          enum:
            - OMI_REALTIME
            - OMI_MEMORY
            - SIMULATED
        kind:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - dedupeKey
        - sessionId
        - text
        - speaker
        - isUser
        - source
        - kind
        - createdAt
      additionalProperties: false
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: unsaid_session
      description: From POST /auth/login
    bearerAuth:
      type: http
      scheme: bearer
      description: API_KEY, for scripts
    webhookSecret:
      type: apiKey
      in: query
      name: secret
    webhookSecretHeader:
      type: apiKey
      in: header
      name: x-omi-secret